set up emailjs and styling
