+++ src/styles/AddItemButton.css
@@ -1,0 +1,36 @@
+.email-field {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 8px;
+}
+
+/* .email-field label {
+  font-family: "Roboto";
+  font-weight: 500;
+  font-size: 16px;
+  line-height: 20px;
+} */
+
+.email-field input {
+  width: 100%;
+  height: 40px;
+  padding: 8px;
+  border: 1px solid #006d7d;
+  border-radius: 4px;
+  font-family: "Roboto";
+  font-weight: 500;
+  font-size: 14px;
+}
+
+.email-field input:focus {
+  outline: none;
+  box-shadow: 0 0 3px rgba(0, 109, 125, 0.5);
+}
+
+.email-field input::placeholder {
+  font-family: "Roboto";
+  font-weight: 500;
+  font-size: 14px;
+  color: #9e9e9e;
+}
